\name{easyrodbcexcel-package}
\alias{easyrodbcexcel-package}
\docType{package}
\title{Simple database interface for RODBC Excel.}
\description{Simple database interface for RODBC Excel. Package build on top of RODBC. Provides functions for reading, writing, listing and subsetting tables in Excel. Allows to retrieve, write and subset Excel files's spreadsheets in a similar way as data.frame's, using single square brackets db[,,,]. easyrodbcexcel internalise many database operations (otherwise tedious) such as rows and columns subsetting (no SQL needed) and connection and disconnection to / from databases.}
\details{
\tabular{ll}{Package: \tab easyrodbcexcel\cr
Version: \tab 0.7.2\cr
Date: \tab 2012-10-22\cr
Title: \tab Simple database interface for RODBC Excel.\cr
Author: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Maintainer: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Depends: \tab R (>= 2.15.1), RODBC, easydb\cr
Suggests: \tab RODBC, easydb\cr
License: \tab AGPL-3\cr
URL: \tab http://easydb.r-forge.r-project.org/\cr}
}
\author{Julien MOEYS <Julien.Moeys@mark.slu.se>}

\keyword{ package }


